feat: added consumer routes

+++ product/productControllers.js
@@ -1,15 +1,10 @@
 const multer = require("multer");
 const multerS3 = require("multer-s3-transform");
-const dotenv = require("dotenv");
 const sharp = require("sharp");
 const aws = require("aws-sdk");
 const { nanoid } = require("nanoid");
 const productServices = require("./productServices");
-const jwt = require("jsonwebtoken");
-const adminRoleModel = require("../admin/adminRoleModel");
-const {
-  checkPermissions,
-} = require("../admin/adminServices/adminAuthServices");
+const adminAuthServices = require("../admin/adminServices/adminAuthServices");
 
 const spacesEndpoint = new aws.Endpoint(process.env.PRIVATE_BUCKET_URL);
 const s3 = new aws.S3({
@@ -24,7 +19,6 @@ const upload = multer({
     bucket: process.env.PUBLIC_BUCKET_NAME,
     acl: "public-read",
     shouldTransform: function (req, file, cb) {
-      // req.body.Id = nanoid(7); //=> "5-JDFkc"
       cb(null, /^image/i.test(file.mimetype));
     },
     transforms: [
@@ -45,8 +39,9 @@ const upload = multer({
 
 const createProduct = async (req, res, next) => {
   try {
-    let checkPermissions = await checkPermissions(req.body.adminToken);
-    console.log(checkPermissions);
+    let checkPermissions = await adminAuthServices.checkPermissions(
+      req.body.adminToken
+    );
     if (checkPermissions.createProduct) {
       const imagesResult = [];
       if (req.files) {
@@ -130,12 +125,14 @@ const readAllProductsAdmin = async (req, res) => {
 
 const updateProduct = async (req, res) => {
   try {
-    let checkPermissions = await checkPermissions(req.body.adminToken);
-    console.log(checkPermissions);
+    let checkPermissions = await adminAuthServices.checkPermissions(
+      req.body.adminToken
+    );
     if (checkPermissions.createProduct) {
       const productsVariants = JSON.parse(req.body.variants);
       let newResult = [];
       const previousImg = req.body.images.split(" ");
+      console.log(req.body.images);
       if (
         previousImg !== [] &&
         previousImg !== " " &&
@@ -211,115 +208,125 @@ const updateProduct = async (req, res) => {
 
 const updateVariants = async (req, res) => {
   try {
-    const product = { _id: req.params.id };
-    const productToUpdate = await productServices.readById(product);
-    const productv2 = productToUpdate.products[0];
+    let checkPermissions = await adminAuthServices.checkPermissions(
+      req.body.adminToken
+    );
+    if (checkPermissions.createProduct) {
+      const product = { _id: req.params.id };
+      const productToUpdate = await productServices.readById(product);
+      const productv2 = productToUpdate.products[0];
 
-    const newVariant = {
-      _id: req.body.variant_id,
-      variantImage: [],
-      active: Boolean(req.body.variantActive),
-      name: req.body.variantName,
-      description: req.body.variantDescription,
-      category: req.body.variantCategory,
-      considerations: req.body.variantConsiderations,
-      attributes:
-        typeof req.body.attributesName0 === "string" &&
-        typeof req.body.attributesName1 === "string"
-          ? [
-              {
-                name: req.body.attributesName0,
-                value: req.body.attributesValue0,
-              },
-              {
-                name: req.body.attributesName1,
-                value: req.body.attributesValue1,
-              },
-            ]
-          : typeof req.body.attributesName0 === "string"
-          ? [
-              {
-                name: req.body.attributesName0,
-                value: req.body.attributesValue0,
-              },
-            ]
-          : [],
-      publicPrice: {
-        from: req.body.variantPublicPriceFrom,
-        to: req.body.variantPublicPriceTo,
-        equation: req.body.variantPublicPriceEq,
-      },
-      prixerPrice: {
-        from: req.body.variantPrixerPriceFrom,
-        to: req.body.variantPrixerPriceTo,
-        equation: req.body.variantPrixerPriceEq,
-      },
-    };
-    const previousImg = req.body.images.split(" ");
-    if (
-      previousImg &&
-      typeof previousImg === "string" &&
-      previousImg.includes("variantImage")
-    ) {
-      newVariant.variantImage.push({
-        type: "images",
-        url: previousImg,
-      });
-    } else if (previousImg && previousImg.length > 1) {
-      previousImg.map((img) => {
-        img.includes("variantImage") &&
-          newVariant.variantImage.push({
-            type: "images",
-            url: img.trim().replace(/[,]/gi, ""),
-          });
-      });
-    }
-    if (req.body.video !== undefined && req.body.video !== "undefined") {
-      newVariant.variantImage.push({
-        type: "video",
-        url: req.body.video,
-      });
-    }
-    if (req.files !== undefined) {
-      req.files.map((img, i) => {
+      const newVariant = {
+        _id: req.body.variant_id,
+        variantImage: [],
+        active: Boolean(req.body.variantActive),
+        name: req.body.variantName,
+        description: req.body.variantDescription,
+        category: req.body.variantCategory,
+        considerations: req.body.variantConsiderations,
+        attributes:
+          typeof req.body.attributesName0 === "string" &&
+          typeof req.body.attributesName1 === "string"
+            ? [
+                {
+                  name: req.body.attributesName0,
+                  value: req.body.attributesValue0,
+                },
+                {
+                  name: req.body.attributesName1,
+                  value: req.body.attributesValue1,
+                },
+              ]
+            : typeof req.body.attributesName0 === "string"
+            ? [
+                {
+                  name: req.body.attributesName0,
+                  value: req.body.attributesValue0,
+                },
+              ]
+            : [],
+        publicPrice: {
+          from: req.body.variantPublicPriceFrom,
+          to: req.body.variantPublicPriceTo,
+          equation: req.body.variantPublicPriceEq,
+        },
+        prixerPrice: {
+          from: req.body.variantPrixerPriceFrom,
+          to: req.body.variantPrixerPriceTo,
+          equation: req.body.variantPrixerPriceEq,
+        },
+      };
+      const previousImg = req.body.images.split(" ");
+      if (
+        previousImg &&
+        typeof previousImg === "string" &&
+        previousImg.includes("variantImage")
+      ) {
         newVariant.variantImage.push({
           type: "images",
-          url: img.transforms[0].location,
+          url: previousImg,
         });
-      });
-    }
-    if (typeof req.body.attributesName === "object") {
-      const a = req.body.attributesName.map((name, i) => {
-        const b = req.body.attributesValue.map((value) => {
-          return value;
+      } else if (previousImg && previousImg.length > 1) {
+        previousImg.map((img) => {
+          img.includes("variantImage") &&
+            newVariant.variantImage.push({
+              type: "images",
+              url: img.trim().replace(/[,]/gi, ""),
+            });
         });
-        return {
-          name: name,
-          value: b[i],
-        };
+      }
+      if (req.body.video !== undefined && req.body.video !== "undefined") {
+        newVariant.variantImage.push({
+          type: "video",
+          url: req.body.video,
+        });
+      }
+      if (req.files !== undefined) {
+        req.files.map((img, i) => {
+          newVariant.variantImage.push({
+            type: "images",
+            url: img.transforms[0].location,
+          });
+        });
+      }
+      if (typeof req.body.attributesName === "object") {
+        const a = req.body.attributesName.map((name, i) => {
+          const b = req.body.attributesValue.map((value) => {
+            return value;
+          });
+          return {
+            name: name,
+            value: b[i],
+          };
+        });
+        newVariant.attributes.push(a);
+      }
+
+      if (productv2.variants !== []) {
+        productv2.variants.map((variant, i) => {
+          if (variant._id === newVariant._id) {
+            productv2.variants.splice(i, 1);
+          }
+        });
+      }
+
+      productv2.variants.push(newVariant);
+
+      const productResult = await productServices.updateProduct(
+        productv2,
+        req.params.id
+      );
+      data = {
+        productResult,
+        success: true,
+      };
+      return res.send(data);
+    } else {
+      return res.send({
+        success: false,
+        message: "No tienes autorización para realizar esta acción.",
       });
-      newVariant.attributes.push(a);
     }
-
-    if (productv2.variants !== []) {
-      productv2.variants.map((variant, i) => {
-        if (variant._id === newVariant._id) {
-          productv2.variants.splice(i, 1);
-        }
-      });
-    }
-
-    productv2.variants.push(newVariant);
-
-    const productResult = await productServices.updateProduct(
-      productv2,
-      req.params.id
-    );
-    data = {
-      productResult,
-      success: true,
-    };
-    return res.send(data);
   } catch (err) {
     console.log(err);
     res.status(500).send(err);
@@ -327,19 +334,40 @@ const updateVariants = async (req, res) => {
 };
 
 async function deleteProduct(req, res) {
-  const productResult = await productServices.deleteProduct(req);
-  data = {
-    productResult,
-    success: true,
-  };
-  return res.send(data);
+  let checkPermissions = await adminAuthServices.checkPermissions(
+    req.body.adminToken
+  );
+  if (checkPermissions.createProduct) {
+    const productResult = await productServices.deleteProduct(req);
+    data = {
+      productResult,
+      success: true,
+    };
+    return res.send(data);
+  } else {
+    return res.send({
+      success: false,
+      message: "No tienes autorización para realizar esta acción.",
+    });
+  }
 }
 
 async function deleteVariant(req, res) {
-  const variantToDelete = await productServices.deleteVariant(req.body);
-  data = { variantToDelete, success: true };
-  return res.send(data);
+  let checkPermissions = await adminAuthServices.checkPermissions(
+    req.body.adminToken
+  );
+  if (checkPermissions.createProduct) {
+    const variantToDelete = await productServices.deleteVariant(req.body);
+    data = { variantToDelete, success: true };
+    return res.send(data);
+  } else {
+    return res.send({
+      success: false,
+      message: "No tienes autorización para realizar esta acción.",
+    });
+  }
 }
+
 module.exports = {
   upload,
   createProduct,
